# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Accounting System",
    "description": "Personal finance tracking with bank statement imports",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/statements/upload": {
      "post": {
        "tags": [
          "statements"
        ],
        "summary": "Upload Statement",
        "description": "Upload and import a bank statement (PDF or CSV).",
        "operationId": "upload_statement_api_statements_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_statement_api_statements_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/statements/{statement_id}": {
      "get": {
        "tags": [
          "statements"
        ],
        "summary": "Get Statement",
        "description": "Get a single statement by ID.",
        "operationId": "get_statement_api_statements__statement_id__get",
        "parameters": [
          {
            "name": "statement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Statement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "description": "List all classification rules.",
        "operationId": "list_rules_api_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  },
                  "type": "array",
                  "title": "Response List Rules Api Rules Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Rule",
        "description": "Create a new classification rule.",
        "operationId": "create_rule_api_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/{rule_id}": {
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Delete Rule",
        "description": "Delete a rule.",
        "operationId": "delete_rule_api_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/{rule_id}/toggle": {
      "patch": {
        "tags": [
          "rules"
        ],
        "summary": "Toggle Rule",
        "description": "Toggle rule active status.",
        "operationId": "toggle_rule_api_rules__rule_id__toggle_patch",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_statement_api_statements_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_statement_api_statements_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "statement_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement Id"
          },
          "transactions_imported": {
            "type": "integer",
            "title": "Transactions Imported",
            "default": 0
          },
          "transactions_skipped": {
            "type": "integer",
            "title": "Transactions Skipped",
            "default": 0
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ImportResult",
        "description": "Result of importing a statement."
      },
...
      "RuleCreate": {
        "properties": {
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "pattern_type": {
            "type": "string",
            "title": "Pattern Type",
            "default": "contains"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "created_from_txn_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created From Txn Id"
          }
        },
        "type": "object",
        "required": [
          "pattern",
          "category_id"
        ],
        "title": "RuleCreate",
        "description": "Fields for creating a rule."
      },
      "RuleResponse": {
        "properties": {
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "pattern_type": {
            "type": "string",
            "title": "Pattern Type",
            "default": "contains"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "pattern",
          "category_id",
          "id",
          "created_at"
        ],
        "title": "RuleResponse",
        "description": "Rule response."
      },
...
      "StatementResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_hash": {
            "type": "string",
            "title": "File Hash"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "opening_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opening Balance"
          },
          "closing_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Balance"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "filename",
          "file_hash",
          "period_start",
          "period_end",
          "opening_balance",
          "closing_balance",
          "imported_at"
        ],
        "title": "StatementResponse",
        "description": "Statement response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}